Move Generator to co_generator.hpp
