+++ include/graph/views/co_bfs.hpp
@@ -1,12 +1,11 @@
 #pragma once
 
-#include "vertexlist.hpp"
 #include <variant>
-#include <coroutine>
 
 #include "graph/graph.hpp"
-#include "graph/views/vertexlist.hpp"
 #include "graph/views/incidence.hpp"
+#include "graph/detail/co_generator.hpp"
+
 #include <queue>
 #include <algorithm>
 
@@ -35,143 +34,6 @@ namespace std::graph {
 template <class G, class F>
 concept edge_weight_function = // e.g. weight(uv)
       copy_constructible<F> && is_arithmetic_v<invoke_result_t<F, edge_reference_t<G>>>;
-
-
-// A simple generator from https://en.cppreference.com/w/cpp/language/coroutines.
-template <typename T>
-struct Generator {
-  // The class name 'Generator' is our choice and it is not required for coroutine
-  // magic. Compiler recognizes coroutine by the presence of 'co_yield' keyword.
-  // You can use name 'MyGenerator' (or any other name) instead as long as you include
-  // nested struct promise_type with 'MyGenerator get_return_object()' method.
-
-  struct promise_type; // REQUIRED type name
-  using handle_type = std::coroutine_handle<promise_type>;
-
-  struct promise_type // required
-  {
-    T                  value_;
-    std::exception_ptr exception_;
-
-    Generator           get_return_object() { return Generator(handle_type::from_promise(*this)); }
-    std::suspend_always initial_suspend() { return {}; }
-    std::suspend_always final_suspend() noexcept { return {}; }
-    void                unhandled_exception() { exception_ = std::current_exception(); } // saving
-    // exception
-
-    template <std::convertible_to<T> From> // C++20 concept
-    std::suspend_always yield_value(From&& from) {
-      value_ = std::forward<From>(from); // caching the result in promise
-      return {};
-    }
-    void return_void() {}
-  };
-
-  handle_type h_;
-
-  Generator(handle_type h)
-        : h_(h) //
-  {}
-  ~Generator() //
-  {
-    h_.destroy();
-  }
-  explicit operator bool() {
-    fill(); // The only way to reliably find out whether or not we finished coroutine,
-            // whether or not there is going to be a next value generated (co_yield)
-            // in coroutine via C++ getter (operator () below) is to execute/resume
-            // coroutine until the next co_yield point (or let it fall off end).
-            // Then we store/cache result in promise to allow getter (operator() below
-            // to grab it without executing coroutine).
-    return !h_.done();
-  }
-  T operator()() {
-    fill();
-    full_ = false; // we are going to move out previously cached
-                   // result to make promise empty again
-    return std::move(h_.promise().value_);
-  }
-
-private:
-  bool full_ = false;
-
-  void fill() {
-    if (!full_) {
-      h_();
-      if (h_.promise().exception_)
-        std::rethrow_exception(h_.promise().exception_);
-      // propagate coroutine exception in called context
-
-      full_ = true;
-    }
-  }
-};
-
-#  if 0
-Generator<std::uint64_t> fibonacci_sequence(unsigned n) {
-  if (n == 0)
-    co_return;
-
-  if (n > 94)
-    throw std::runtime_error("Too big Fibonacci sequence. Elements would overflow.");
-
-  co_yield 0;
-
-  if (n == 1)
-    co_return;
-
-  co_yield 1;
-
-  if (n == 2)
-    co_return;
-
-  std::uint64_t a = 0;
-  std::uint64_t b = 1;
-
-  for (unsigned i = 2; i < n; ++i) {
-    std::uint64_t s = a + b;
-    co_yield s;
-    a = b;
-    b = s;
-  }
-}
-
-class fib_seq {
-public:
-  fib_seq(unsigned n) : n_(n) {}
-
-  Generator<std::uint64_t> operator()() {
-    if (n_ == 0)
-      co_return;
-
-    if (n_ > 94)
-      throw std::runtime_error("Too big Fibonacci sequence. Elements would overflow.");
-
-    co_yield 0;
-
-    if (n_ == 1)
-      co_return;
-
-    co_yield 1;
-
-    if (n_ == 2)
-      co_return;
-
-    std::uint64_t a = 0;
-    std::uint64_t b = 1;
-
-    for (unsigned i = 2; i < n_; ++i) {
-      std::uint64_t s = a + b;
-      co_yield s;
-      a = b;
-      b = s;
-    }
-  }
-
-private:
-  unsigned n_ = 0;
-};
-#  endif //0
 
 
 // These events duplicate boost::graph's BFSVisitorConcept
@@ -292,6 +154,9 @@ Generator<bfs_value_t<G>> co_bfs(G&&            g,    // graph
 
 
 enum class dijkstra_event : int {
+  //tree_edge,
+  //gray_target,
+
   initialize_vertex,
   discover_vertex,
   examine_vertex,
@@ -301,6 +166,27 @@ enum class dijkstra_event : int {
   finish_vertex
 };
 
+// BGL list of all possible events?
+enum event_visitor_enum {
+  on_no_event_num,
+  on_initialize_vertex_num,
+  on_start_vertex_num,
+  on_discover_vertex_num,
+  on_finish_vertex_num,
+  on_examine_vertex_num,
+  on_examine_edge_num,
+  on_tree_edge_num,
+  on_non_tree_edge_num,
+  on_gray_target_num,
+  on_black_target_num,
+  on_forward_or_cross_edge_num,
+  on_back_edge_num,
+  on_finish_edge_num,
+  on_edge_relaxed_num,
+  on_edge_not_relaxed_num,
+  on_edge_minimized_num,
+  on_edge_not_minimized_num
+};
 
 } // namespace std::graph
